requirements.txt
pillow
mss==9.0.1
numpy==1.26.4 